✨ Support wallet direct claim for free nft book
